requirements.txt
python-dotenv
ffmpeg-python
PyNaCl
audioop-lts
discord
pillow
faker